Refactor event data transfer objects (DTOs) and update imports
