+++ docs/designs/game-flow.png
