+++ src/controllers/game.controller.ts
@@ -1,10 +1,6 @@
 import { Socket } from "socket.io";
-import { StartGameData, MakeMoveData, LeaveGameData } from "../dtos";
-
-function startGame(socket: Socket, data: StartGameData) {}
+import { MakeMoveData } from "../types/dtos";
 
 function makeMove(socket: Socket, data: MakeMoveData) {}
 
-function leaveGame(socket: Socket, data: LeaveGameData) {}
-
-export { startGame, makeMove, leaveGame };
+export { makeMove };
